+++ Plugin/src/Dependency/LethalLevelLoaderProxy.cs
@@ -1,9 +1,8 @@
-using System;
-using System.Collections.Generic;
 using System.Runtime.CompilerServices;
 using BepInEx.Bootstrap;
-using JetBrains.Annotations;
 using LethalLevelLoader;
+using MattyFixes.Interfaces;
+using MattyFixes.Utils;
 
 namespace MattyFixes.Dependency;
 
@@ -19,18 +18,29 @@ public static class LethalLevelLoaderProxy
             return _enabled.Value;
         }
     }
-
+    
+    // ReSharper disable function SuspiciousTypeConversion.Global
     [MethodImpl(MethodImplOptions.NoInlining | MethodImplOptions.NoOptimization)]
-    public static void GetModdedItems([NotNull] in Dictionary<Item, (string api, string modname)> items)
+    public static void PopulateModdedItems()
     {
         
         MattyFixes.Log.LogWarning("LethalLevelLoader found, reading PatchedContent.ExtendedItems");
         foreach (var extendedItem in PatchedContent.ExtendedItems)
         {
-            if (extendedItem.ContentType == ContentType.Vanilla)
+            if (((IInjectedItem)extendedItem.Item).MattyFixes_ItemType >= ItemCategory.ItemType.Modded)
                 continue;
 
-            items.TryAdd(extendedItem.Item, ("LethalLevelLoader", extendedItem.ModName));
+            if (extendedItem.ContentType == ContentType.Vanilla)
+            {
+                ((IInjectedItem)extendedItem.Item).MattyFixes_ItemType = ItemCategory.ItemType.Vanilla;
+                ((IInjectedItem)extendedItem.Item).MattyFixes_Path = extendedItem.Item.ComputePath("Vanilla");
+            }
+            else
+            {
+                ((IInjectedItem)extendedItem.Item).MattyFixes_ItemType = ItemCategory.ItemType.Modded;
+                ((IInjectedItem)extendedItem.Item).MattyFixes_Path =
+                    extendedItem.Item.ComputePath("LethalLevelLoader", extendedItem.ModName);
+            }
         }
     }
 }